--- src/ar/edu/unq/po2/zonaOrganizacionUbicacion/Organizacion.java
@@ -1,5 +1,0 @@
-package ar.edu.unq.po2.zonaOrganizacionUbicacion;
-
-public class Organizacion {
-
-}
