+++ src/web_vinchucas/Organizacion.java
@@ -1,0 +1,5 @@
+package web_vinchucas;
+
+public class Organizacion {
+
+}
