+++ src/web_vinchucas/Ubicación.java
@@ -1,0 +1,5 @@
+package web_vinchucas;
+
+public class Ubicación {
+
+}
