+++ src/web_vinchucas/ZonaCobertura.java
@@ -1,0 +1,5 @@
+package web_vinchucas;
+
+public class ZonaCobertura {
+
+}
